API 오류수정

+++ src/PAGES/DOCUMENT/TempDocumentDetail.jsx
@@ -6,44 +6,55 @@ import {
 import {useNavigate, useParams} from "react-router-dom";
 import {Button} from "react-bootstrap";
 import styles from "../APPROVAL/Write.module.css"
-import {Viewer} from "@toast-ui/react-editor";
-import '@toast-ui/editor/dist/toastui-editor-viewer.css';
 import TempDocumentSignTable from "./components/TempDocumentSignTable";
 import Swal from "sweetalert2";
+import parse from "html-react-parser";
 
-function DocumentDetail() {
+function TempDocumentDetail() {
     const {id} = useParams();
     const navigate = useNavigate();
     const [documentData, setDocumentData] = useState({});
     const [signLine, setSignLine] = useState([]);
     const [isCompleted, setIsCompleted] = useState(false);
 
-
     useEffect(() => {
         fetcher.get(`${TEMP_DOCUMENT_READ_API}/${id}`)
             .then((res) => {
-                const { document, groupedApprovals } = res.data
+                // const { document, groupedApprovals } = res.data
                 // 문서 정보
-                setDocumentData(document)
-                // 결재라인
-                setSignLine(groupedApprovals[document.sno])
+                setDocumentData(res.data)
+                // 결재라인 - 임시저장에서 결재라인 아직 구현안됨
+                // setSignLine(groupedApprovals[document.sno])
                 setIsCompleted(true)
             })
+            .catch((err)=>console.log(err))
     }, [id])
-    // TODO: API 오류 수정중
+
 
     const deleteBtn = () => {
-        fetcher.delete(`${DOCUMENT_DELETE_API}/${id}`)
-            .then(
-                Swal.fire({
-                    position: 'mid',
-                    icon: 'success',
-                    title: '삭제 완료',
-                    showConfirmButton: false,
-                    timer: 1500
-                }),
-                navigate(-1)
-            )
+        Swal.fire({
+            title: "문서를 삭제 하시겠습니까?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: '확인',
+            cancelButtonText: '취소'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetcher.delete(`${DOCUMENT_DELETE_API}/${id}`)
+                    .then(
+                        Swal.fire({
+                            position: 'mid',
+                            icon: 'success',
+                            title: '삭제 완료',
+                            showConfirmButton: false,
+                            timer: 1500
+                        }),
+                        navigate(-1)
+                    )
+            }
+        })
     }
 
     return (
@@ -51,7 +62,7 @@ function DocumentDetail() {
             <div className={styles.upperContainer}>
                 <div className={styles.buttonGroup}>
                     <Button className="button" onClick={()=>navigate(-1)}>목록으로</Button>
-                    <Button className="button">상신하기</Button>
+                    <Button className="button">문서수정</Button>
                     <Button className="button" onClick={deleteBtn}>문서삭제</Button>
                 </div>
             </div>
@@ -68,13 +79,13 @@ function DocumentDetail() {
                     <div className={styles.documentTitle}>
                         <p>제목 : </p>{documentData.title}
                     </div>
-                    {/*<div className={styles.documentContent}>*/}
-                    {/*    {isCompleted && <Viewer initialValue={documentData.content}/>}*/}
-                    {/*</div>*/}
+                    <div>
+                        <div>{isCompleted && parse(documentData.content)}</div>
+                    </div>
                 </div>
             </div>
         </div>
     )
 }
 
-export default DocumentDetail
+export default TempDocumentDetail
